Add delete btn and change read to status
Added a delete button icon remove books from library. Not yet
functional. Positioned the button in the bottom right corner for
better user experience and more universal location.
Changed how the read, not read, reading status was displayed for
better noticeability and make it clearer what it was communicating.
Added a demo book card in the html to know what elements to add in js
and update css layout.